server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
torch==2.4.0
torchvision==0.19.0
hloc @ git+https://github.com/cvg/Hierarchical-Localization.git
pycolmap==0.6.1
